kafka_faust_app/requirements.txt
aiohttp==3.6.2
aiohttp-cors==0.7.0
async-timeout==3.0.1
attrs==19.3.0
cassandra-driver==3.23.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
colorclass==2.2.0
colorlog==4.1.0
croniter==0.3.31
faust==1.10.4
geomet==0.1.2
idna==2.9
kafka-python==1.4.7
mode==4.3.2
multidict==4.7.5
mypy-extensions==0.4.3
opentracing==1.3.0
python-dateutil==2.8.1
robinhood-aiokafka==1.1.6
six==1.14.0
terminaltables==3.1.0
typing-extensions==3.7.4.2
venusian==1.2.0
yarl==1.4.2
